Refactored to shift our logic to mod.rs
